Remove jsgbemu and link to web resources.
The web/ directory containing `jsgbemu` dominated the git archive size.
This should really be a git submodule or something. It also seems to be
fairly buggy, in practice button input doesn't seem to be working. It's
too early to sped time fixing web junk for an unfinished game so let's
remove it until we're ready.